Add manylinux2014 migration to deprecations page (#5459)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [ ] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** As Centos7 is EOL, we will be migrating Lightning and
Catalyst to manylinux_2_28. This will drop support for older linux
variants that do not have a modern enough library ecosystem to maintain
compatibility. This PR updates the deprecations page to ensure this is
known.

**Description of the Change:** Notify end-users of migration from
manylinux2014 to manylinux_2_28

**Benefits:** Allows developers on affected systems to know in advance
if their OS needs to be updated.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>