add vmap support for devices that do not support parameter broadcasting (#5286)
Fixes #5240 [sc-57137] [sc-57848] Fixes #5289

Basically when we set `vectorized=True` inside the `pure_callback` call,
we assumed that the device natively supports broadcasting. And then we
only tested with devices that did indeed natively support parameter
broadcasting.

This problem was made worse by the fact that our `vmap` tests included a
`Hamiltonian` expectation value, which caused up to skip many of the
test cases that we really should have been testing. So I got rid of the
`Hamiltonian` from the test so we could actually test more situations.

I also added more `lightning.qubit` tests to the test configuration.
That forced one or two other changes.

The major problem with `jax.vmap` is that it adds in a parameter
broadcasting dimension *after* we have already handled all of our
preprocessing and breaking up parameter broadcasting.

---------

Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: Nathan Killoran <co9olguy@users.noreply.github.com>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>
Co-authored-by: Mikhail Andrenkov <mikhail@xanadu.ai>
Co-authored-by: Korbinian Kottmann <43949391+Qottmann@users.noreply.github.com>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <isaacdevlugt@gmail.com>
Co-authored-by: Pietropaolo Frisoni <pietropaolo.frisoni@xanadu.ai>
Co-authored-by: soranjh <40344468+soranjh@users.noreply.github.com>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>