[bugfix] Probabilities do not sum to one with Torch (#5462)
**Context:** When computing the `expval` of an operator using a quantum
device with the `torch` interface and `default_dtype` set to
`torch.float32`, the probabilities do not sum to one. This error does
not occur if `default_dtype` is set to `torch.float64`.

**Description of the Change:** A renormalization of probabilities is
introduced to overcome the issue. Such renormalization of probabilities
occurs whenever the following two conditions are satisfied: 1) There is
at least one probability that does not sum precisely to one, and 2) Such
difference for all probabilities is between `0` and `1e-07`.

**Benefits:** The error is not raised anymore.

**Possible Drawbacks:** The main drawback is that renormalization can
occur in cases where it should not happen, although it is unlikely since
the cutoff `1e-07` is expected to be reasonably small to prevent such
cases (but large enough not to raise the error anymore).

**Related GitHub Issues:** #5444

[sc-59957]