add type hint for unimplemented operator function (#5490)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
The abstract class Operator defines several methods that need to be
implemented in subclasses. While these methods are defined to raise
errors before implementation in the abstract class, this leads to
linting errors. Therefore, appropriate type hints were added after
implementation.

**Description of the Change:**
Add type hints to the following methods.
- `compute_matrix`
- `matrix`
- `compute_sparse_matrix`
- `sparse_matrix`
- `compute_eigvals`
- `decomposition`
- `compute_decomposition`
- `compute_diagonalizing_gates`
- `single_qubit_rot_angles`
- `compute_kraus_matrices`

**Benefits:**
Improve the interpretability of the linter

**Possible Drawbacks:**

**Related GitHub Issues:**
Add type hints to pennylane.operation.Operator #5398

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>