Update device tests to work with Lightning devices (#5518)
The new tests added to the device test suite weren't totally compatible
with lightning devices. I've updated the failing tests to work now. The
tensorflow tests failed because our execution pipeline with tensorflow
doesn't correctly respect the dtypes, which is a larger issue for us to
resolve, so it will be handled in a different PR. For now I've skipped
the failing tests.