Introduce special diagonalize procedure around ProbabilityMP measurements for Lightning  (#5529)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
`_qubit_device`'s `statistics` only passes the
[wires](https://github.com/PennyLaneAI/pennylane/blob/master/pennylane/_qubit_device.py#L672C17-L672C99)
argument to `probability`. When this is called on the Lightning side,
the basis info is not there to diagonalize the state. This does not
happen in LQ since the whole measurement is passed to `probs`.
`expval/var` however accept `obs` which allows old device like
L-Kokkos/GPU to diagonalize the state in these methods.
 
**Description of the Change:**
To avoid diagonalizing twice for `expval/var`, we
diagonalize/undiagonalize the state around in `ProbabilityMP`
measurements in `statistics` when executing on a Lightning device.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>