Allow gradient transforms applied to broadcasted tapes if the broadcasted parameter(s) are not trainable (#5452)
**Context:**
In #4462 we discussed the differentiation of broadcasted tapes/QNodes
with gradient transforms. Due to some technical debt in the code for
JVPs and due to unclear desired behaviour, applying gradient transforms
to such QNodes was disallowed in #4480.
However, this also disallows applying gradient transforms to broadcasted
tapes for which the broadcasted parameter(s) is/are not trainable, even
though the problems discussed in #4462 do not apply in this case.

**Description of the Change:**
This PR lifts the restriction of applying gradient transforms to
broadcasted tapes for the case that no trainable parameters are
broadcasted.
The problem described in [this forum
post](https://discuss.pennylane.ai/t/qml-training-using-shot-based-simulator/4206)
can be resolved by this, for example.

**Benefits:**
Unlocks gradient transforms for broadcasted tapes/QNodes as long as the
broadcasting is in non-trainable parameters (which is quite common for
QML applications: You want to broadcast across data points rather than
trainable parameters.).

**Possible Drawbacks:**
Usage of `tape.trainable_params` although we don't want to rely on this
property too much in the future.