Add test workflow for legacy opmath (#5435)
Currently, the tests only run with the new opmath, so won't know if we
introduced a bug that breaks the legacy opmath behaviour while it is
still in its deprecation cycle.

This PR creates a kwarg, `disable-opmath`, that can be passed when
running the tests, and a corresponding workflow that runs tests with
that kwarg set to True. This can also be used locally, i.e. `python -m
pytest tests/ --disable-opmath=True`

Right now, it runs with CI on the PR, but once the tests are all
passing, the line triggering that will be removed and it will only run
every 3-4 days, in the middle of the night. Then we will add it to the
test matrix (separate PR to modify the plugin test matrix repo).

The changes to .yml files and to the `conftest` files are all about
allowing us to run these additional tests. A few modifications to tests
were made to allow them to pass with both legacy opmath and new opmath.

There is one test currently marked as xfail for new opmath that I would
call a bug - I opened an issue here:
https://github.com/PennyLaneAI/pennylane/issues/5512

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: qottmann <korbinian.kottmann@gmail.com>