Default qutrit mixed preprocess (#5451)
**Context:**
This PR begins instantiating a qutrit mixed device based on the new
qml.devices.Device API. It is focused on the preprocess method
instantiation. This is a continuation of other additions I have made to
build a qutrit mixed state device.

**Description of the Change:**
This PR adds default_qutrit_mixed device, adding the class, it's
initialization, and an instantiation of the preprocess method. The
instantiation includes a stub for execute.

**Benefits:**
Necessary to continue in the addition of a qutrit mixed state device.

**Possible Drawbacks:**
Execute has not been added yet.

**Related GitHub Issues:**
N/A

---------

Co-authored-by: Olivia Di Matteo <2068515+glassnotes@users.noreply.github.com>
Co-authored-by: Gabriel Bottrill <bottrill@student.ubc.ca>