Split prng_keys with JAX. (#5515)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
JAX RNG keys should not be reused (in production) according to the JAX
documentation. The current usage only split the keys at the lowest level
(`_sample_state_jax`). This could cause unwanted correlations. For
example, we reuse the device key for each tape in a batch of tapes in
`default.qubit`.
 
**Description of the Change:**
Add functions and methods to split RNG keys. Split RNG keys to avoid
reusing keys.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
#5337 

[sc-61435]