[DLA 1] Add qml.dla.VSpace class for keeping track of linearly independent vectors (#5161)
This class enables easy implementation for `qml.dla.lie_closure`
realized in the follow-up PR.
The class VSpace is currently not (yet) user-facing and acts as a
utility to lie_closure. We could consider making it user-facing though.
Open to discussing this! The main reason I am holding back is that the
UI is experimental and some things may be subject to change /
improvement, so I dont want to lock in just yet.

The main point of this class is to efficiently keep track of linearly
independent operators that span a vector space, specifically for
PauliSentences.

This is done by storing and processing ``M``, which
is a dictionary-of-keys (DOK) style sparse representation of the set of
basis vectors. You can
think of it as the numpy-equivalent of a PauliSentence: each
``PauliWord`` (key of ``PauliSentence``)
represents one row of ``M`` with the coefficient (value of
``PauliSentence``).
For example the set of 3 linearly independent generators ``X(0) + X(1),
X(0) + X(2), X(0) + 0.5 * Y(0)``
can be represented as

```
[
    [1, 1, 1],
    [1, 0, 0],
    [0, 1, 0],
    [0, 0, 0.5]
]
```

where each column represents one sentence, and each row represents the
coefficient of the respective word in the sentence.
To make sense of this representation one additionally needs to keep
track of the mapping between keys and rows. In this case we have

```
pw_to_idx = {
    X(0) : 0,
    X(1) : 1,
    X(2) : 2,
    Y(0) : 3
}
``` 

where we have set the numbering based on appearance in the list of
generators. This mapping is in general not unique.

ToDo

- [x] VSpace class init
- [x] VSpace.add() method
- [x] test init
- [x] test VSpace.add()
- [x] add cheaper check for proportionality before performing rank
computation
- [x] add test for util accordingly
- [x] increase test coverage (i.e. add examples)
- [x] check edge cases
- [x] changelog (no changelog since this is private functionality)
- [x] Fix equality check
- [x] Overhaul `VSpace.add` logic (inplace vs not)

Based off @dwierichs 's
[prototype](https://github.com/PennyLaneAI/pennylane/pull/4852)
Follow-up PR adding lie_closure
https://github.com/PennyLaneAI/pennylane/pull/5169
[sc-51430]

https://github.com/PennyLaneAI/pennylane/pull/5161 (DLA1)
https://github.com/PennyLaneAI/pennylane/pull/5169 (DLA2)
https://github.com/PennyLaneAI/pennylane/pull/5406 (DLA3)

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Vincent Michaud-Rioux <vincentm@nanoacademic.com>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>