Use SampleMPs in dynamic_one_shot (#5486)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
The native MCM workflow breaks the device API where a sequence of
MeasurementProcess objects is expected in the output.

**Description of the Change:**
Introduce SampleMPs in the auxiliary tape. Pass the mid_measurements
dictionary around simulate and sampling to return the correct sample
measurements. Modify the dynamic_one_shot transform post-processing
function accordingly.

**Benefits:**
Conform to current API.
Road to jax.jit support.

**Possible Drawbacks:**
Ad hoc post-processing required in measure_with_samples.

**Related GitHub Issues:**
[sc-60945]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>