Updates to v0.36 release notes (#5554)
**Context:**

The release notes need to be tidied in time for the 0.36 release on May
7th.

**Description of the Change:**

This is the first phase of changes to the release notes. In this PR we
add categories to major features and improvements and categorize the
existing PRs.

**Recommendation for reviewers:**

- It is important to merge this first changelog update PR relatively
quickly as it often goes out of date with `master` and has merge
conflicts. Hence, we typically do not worry about the changelog entries
themselves, just that the categories are added and we believe entries
are in the right category. Changes to the entries themselves, and
writing polished summaries of major features, will happen in a
subsequent PR.
- There are so many changes that the diff can be hard to follow. It is
easiest to simply view the docs build to check if you are happy.

---------

Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>