Improved handling for Identities on no wires (#5570)
**Context:**

While we have added the ability to have `qml.I()` on no wires to better
reflect the wire-independent nature of the identity and create
operations again from a simplified form, we don't necessarily have
extensive support and coverage for this new style of observable yet.

**Description of the Change:**

This PR extends support for `qml.I()` for `devices.qubit` measurement
pathways `measure` and `measure_with_samples`. It also allows us to
calculate the eigenvalues for composite ops that contain an operator on
no wires. And finally, it adds support for observables with no wires to
`qml.pauli.group_observables`.

Throughout, I assume an observable with no wires commutes with
everything, as `qml.I` does in fact commute with everything.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-61509] Fixes #5532

---------

Co-authored-by: dwierichs <david.wierichs@xanadu.ai>