Make the contraction of quantum and classical Jacobians consistent in `gradient_transform` (#4945)
**Context:**
When applying a `gradient_transform` to a QNode, the QNode wrapper makes
use of the function `_contract_qjac_with_cjac` to contract the quantum
and classical Jacobians with each other.
Here, the classical Jacobian is the Jacobian of classical processing
that happens within the QNode between the QNode input arguments and the
tape-level operation arguments.

For the special case where the classical Jacobian is the identity, there
is a separate logic in `_contract_qjac_with_cjac` that skips the
contraction.
This is a problem because the contraction, even if it does not affect
the numerical values of the output, also implies reformatting of the
quantum Jacobian, by virtue of transposition, casting to tuples,
stacking.
All this is skipped when the classical Jacobian is the identity, which
happens for example for 1D QNode arguments that are not processed any
further but just passed to operations within the QNode. Unfortunately,
that's a very common scenario, making this PR a breaking change with
high visibility.

In addition, we never seem to test certain return type setups _with_
classical processing, so that the contraction is skipped and relevant
cases are not covered.

**Description of the Change:**
This PR removes the special logic for the case where the classical
Jacobian is the identity.
There is an alternative version that still skips the contraction but
performs all the reformatting steps. Given the amount of code
duplication, I do not think we want to go for that version.

In addition, this PR adds the contraction logic for previously uncovered
cases, which are now hit because we do not skip the contraction any
longer.

**Benefits:**
The output of `gradient_transform`s applied to QNodes is
self-consistent.
The output of `gradient_transform`s applied to QNodes without classical
processing is consistent with results from using interface entrypoints.

**Possible Drawbacks:**
Breaking change

**Related GitHub Issues:**
Fixes #4940 

[sc-51888]