Add `rng` and `prng_key` to `get_final_state`, `apply_operation` (#5337)
**Context:**
The seed provided to `DefaultQubit` during initialization is only used
for measurements. However, we now have uses for it while applying
operations, specifically with applying mid-circuit measurements with
`dynamic_one_shot`, and updating the number of shots with postselection
with `defer_measurements`. This PR updates
`devices.qubit.simulate.get_final_state` and
`devices.qubit.apply_operation` to use the RNG/PRNGKey provided during
device initialization for stochastic behaviour.

**Description of the Change:**
* Add `rng` and `prng_key` arguments to `get_final_state` and the
`MidMeasureMP` dispatcher of `apply_operation`. `appy_operation` now
uses the rng/PRNGKey to sample the measurement outcome.
* Add `rng` and `prng_key` arguments to `_postselection_postprocess`.
This function is used for renormalizing the state after a `Projector` is
applied to it, as well as to update the number of shots if provided.
This function now uses the rng/PRNGKey to update shots using the
binomial distribution.
* Added/updated tests. The native MCM tests for `default.qubit` now set
a seed on the device. With this change, I also removed `flaky` from all
the native MCM tests and instead increased the number of shots for the
tests that were still failing.
* Updated `get_final_state` and `apply_operation` to accept
`**execution_kwargs` arguments, which hide `rng`, `prng_key`,
`interface`, and `mid_measurements`. All of these arguments impact
execution outcomes, hence why they were collected. This also means that
all of those arguments are now strictly keyword arguments and can no
longer be passed to the functions as positional arguments.

**Benefits:**
* Random behaviour is more consistent and expected.
* Less random, unrelated failures in CI.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Astral Cai <astral.cai@xanadu.ai>