Improve performance of differentiable matrix calculations for Pauli Sentences (#5578)
**Context:**

The new operator arithmetic now enabled provides a fully differentiable
pathway for calculating the matrix that was not available before.

Unfortunately, a differentiable calculation of a large dense matrix can
be quite slow, and users may not realize they should be using a sparse
matrix formulation instead. Just because we offer it does not make it a
universally good idea.

This PR hopes to close some of that performance gap, by taking advantage
of the structure of pauli sentences to improve the efficiency of
calculating the matrix.

**Description of the Change:**

Adds in a new pathway for calculating the matrix of a pauli sentence.

**Benefits:**

Some performance numbers for the H5 molecule follow:

![Screenshot 2024-04-25 at 9 57
05 AM](https://github.com/PennyLaneAI/pennylane/assets/6364575/14f1211f-8d60-4b72-9c32-a25cb1e08e74)

While the differentiable matrix is still not as fast as sparse matrix
calculation, this method makes up a lot of ground. Note that the dense
matrix is now on the same order of magnitude as the sparse matrix
calculation for the legacy hamiltonian.

This formulation is also jit-able, and jitting the calculation of the
matrix allows us to pick up another order of magnitude improvement.

**Possible Drawbacks:**

Dense is still not as fast as the sparse matrix, and might not scale as
well as the sparse formulation.

Also, this pathway is much less intuitive and involves some numerical
black magic.

**Related GitHub Issues:**

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>