Reduce the number of operations and control wires for `cond` with `defer_measurements` and postselection (#5558)
**Context:**
Deferring measurements implies that we replace operations classically
controlled on mid-circuit measurement values by standard controlled
quantum operations. If postselection is used, we can save on the number
of control wires and on the overall number of operations queued by
`defer_measurements`.

Note that a single postselected measurement can reduce the number of
operations drastically, depending on the classical coprocessing.

**Description of the Change:**
This PR changes `defer_measurements` to only queue operations that are
not reduced to the identity because of postselection rules. In addition,
control wires corresponding to postselected measurements are excluded,
as their value is known at transform time already.

To achieve the above, this PR adds a method `_postselected_items` to
`MeasurementValue` that only returns a (pruned) branch together with the
classical processing result of the non-pruned branch.

**Benefits:**
Reduction of operation count and complexity when using
`defer_measurements` with postselection.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>