Fix `qml.draw/_mpl` to work with deferred measurements circuits that use MCMs in measurements (#5610)
**Context:**
`defer_measurements` leaves `MeasurementValue`s in the terminal
measurement processes if they are present. This is not friendly with the
drawer as the classical wires for those measurement values do not exist
when using `defer_measurements`. This PR adds a "hack" fix so that the
drawer/MPL drawer work with this scenario.

**Description of the Change:**
Add `transform_deferred_measurements_tape` util function, which replaces
any `MeasurementValue`s present in the tape after applying
`defer_measurements` with wires. This transform is only applied in
`tape_text` and `tape_mpl`.

**Benefits:**
Drawer works better with MCMs

**Possible Drawbacks:**
Hacky fix, technical debt to implement cleaner fix later

**Related GitHub Issues:**
https://github.com/PennyLaneAI/pennylane/issues/5588