[Bugfix] Hamiltonian is rendered in top level docs (#5640)
Due to the dispatch of `qml.Hamiltonian` that involves removing
`Hamiltonian` from the top level in `pennylane/__init__.py`, the doc was
not displayed anymore. This fix "hacks" sphinx to add the missing obect
back, at least for sphinx.

Optional separate issue: A user seeking out the `qml.Hamiltonian` docs
at the top level may be confused because this now dispatches to
`LinearCombination` but this is not mentioned there. Added a warning to
avoid that confusion.

---------

Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Astral Cai <astral.cai@xanadu.ai>