dynamic_one_shot is compatible with jax.jit  (#5557)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
As a first step toward our goal to support `dynamic_one_shot` with
Catalyst, we attempt to support `jax.jit`.

**Description of the Change:**
Make all array shapes static. Perform post-selection as a filtering
operation during the post-processing step of `dynamic_one_shot`. Add
`jax.jit` tests. We also introduce the constant `fillin_value` which is
used as a placeholder for postselection mismatched samples.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-61173]
[sc-61315]
[sc-62094]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>